Revert "added a reducer and action to get user's current location, then delet…"
