+++ client/src/actions/types.js
@@ -5,4 +5,3 @@ export const AUTH_ERROR = 'AUTH_ERROR';
 export const ADD_TODO = 'ADD_TODO';
 export const TODO_ERROR = 'TODO_ERROR';
 export const FETCH_TODOS = 'FETCH_TODOS';
-export const FETCH_LOCATION = 'FETCH_LOCATION'
