--- client/src/components/Geoloc.js
@@ -1,30 +1,0 @@
-import React, {Component} from 'react';
-import { fetchLocation } from "./../actions/locationAction"
-import { connect } from 'react-redux';
-import GoogleMap from './GoogleMap'
-
-
-class Geoloc extends Component {
-
-    render() {
-        const position = {
-            lat: this.props.coords && this.props.coords.latitude,
-            lng: this.props.coords && this.props.coords.longitude
-        }
-        if (position.lat && position.lng){
-            this.props.fetchLocation(position);
-        }
-        return (
-            <div>
-                <p>We're Currently using your location to load the map</p>
-                <GoogleMap position = {position}/>
-
-            </div>
-        )
-
-    }
-}
-
-
-export default connect(null, { fetchLocation })(Geoloc);
-
