--- client/src/components/GoogleMap/index.js
@@ -1,51 +1,0 @@
-import React, {Component} from 'react';
-import { Map, GoogleApiWrapper, Marker } from 'google-maps-react';
-
-
-
-export class MapContainer extends Component {
-
-  displayMarkers = () => {
-
-      return <Marker position={{
-        lat: this.props.position.lat,
-        lng: this.props.position.lng
-      }}
-                     onClick={() => console.log("You clicked me!")} />
-
-  }
-
-  render() {
-    const mapStyles = {
-      marginTop: '200px',
-
-
-      width: '70%',
-      height: '70%',
-    };
-    return (
-        <div>
-
-
-          <Map
-              google={this.props.google}
-              zoom={8}
-              style={mapStyles}
-              initialCenter={{ lat: this.props.position.lat, lng: this.props.position.lng}}
-          >
-            {this.displayMarkers()}
-          </Map>
-
-        </div>
-    );
-  }
-}
-
-
-
-export default GoogleApiWrapper({
-  apiKey: 'AIzaSyBAQv_VajKKxKQTr1tw-PpDaMhV30n_D-c'
-})(MapContainer);
-
-
-
