+++ client/src/containers/Home.js
@@ -1,16 +1,15 @@
 import React from 'react';
 // import NavBar from '../components/NavBar';
-// import GoogleMap from '../components/GoogleMap';
+import GoogleMap from '../components/GoogleMap';
 import Search from '../components/Search';
 // import Search from '../components/Search/Search';
 // import NavBar from '../components/NavBar'
-import Demo from '../components/Demo'
 
 export default () => {
   return (
     <div>
       <Search />
-      <Demo />
+      <GoogleMap />
     </div>
   );
 };
