--- client/src/reducers/location.js
@@ -1,13 +1,0 @@
-import { FETCH_LOCATION } from "../actions/types";
-
-
-
-export default function(state = {}, action) {
-
-    if (action.type === FETCH_LOCATION) {
-        console.log("hey there")
-        return {...state, location: action.payload };
-    } else {
-        return state;
-    }
-}
