+++ package.json
@@ -28,7 +28,6 @@
         "passport-jwt": "^4.0.0",
         "passport-local": "^1.0.0",
         "prop-types": "^15.7.2",
-        "react-geolocated": "^2.5.0",
         "validator": "^11.1.0"
     },
     "devDependencies": {
